Add functions.host library
This library is used by osbash host-side scripts.

It contains exec_cmd, a simple function that is used when marking
commands for conditional execution, depending on whether the user
requested Windows batch files and whether the artifacts are to be
created on the local machine as well.

The library also contains functions that allow the host to execute
scripts in the VMs, either directly through ssh or indirectly, by
adding files to an autostart directory shared with the guest VMs.

In addition, the library contains functions to find and (if
necessary) download the install ISO image of the configured
Linux distribution.

Partial-Bug: 1312764
Implements: blueprint openstack-training-labs
Change-Id: I49d49340db89b61a03dbc483291c252c8c72d61a