labs: remove autostart scripts for ssh execution, too
Scripts are copied to the autostart directory before they are executed.
After a script has been executed by osbash through an ssh connection, it
is not removed from that directory, but it should be.

If a VM has both methods enabled (ssh and osbashauto via shared), this
might result in scripts being executed twice.

With this patch, scripts are removed from the autostart directory after
execution, both in the VM and on the host.

Implements: blueprint openstack-training-labs
Change-Id: I74faae2cb302b0faf5865df2922fb00e8a51936e