labs: split wait_for_shutdown off from boot
The boot command in config/scripts.* implies that osbash is going to
wait for the VM to shutdown before it proceeds. That means we currently
can't leave a VM running while booting another which is unfortunate when
the compute node needs access to controller services to build.

This patch splits the boot command into two separate commands, boot and
wait_for_shutdown, so we can omit wait_for_shutdown as desired.

All the configuration files are updated to keep build behavior unchanged,
with one exception: after the controller has been built, shutdown,
and snapshotted, it is booted again.

Change-Id: Ib6ba789f1b41909ef2398e9f78b343c41230e57b
Implements: blueprint openstack-training-labs