labs: fix guest-additions download
If the VirtualBox guest-additions have to be downloaded from the
Internet (usually they are provided by VirtualBox or already on the
disk), osbash fails to find the ISO image immediately -- it needs to
be restarted for that.

With this patch, osbash will correctly detect a freshly downloaded ISO
image.

Change-Id: Id767525cb25a2953e32ca5cba95fedddcdab6a5e
Implements: blueprint openstack-training-labs