labs: add export via clone VM directory
This changeset introduces a new export method: exporting the named node
VMs as clones to the img directory.

The -e option takes now an argument (ova or dir) to distinguish between
exporting to an OVA package file and exporting to the img directory.

Unlike the OVA package files, this method preserves snapshots.

The resulting vbox files can be imported as VMs via VirtualBox GUI
(Machine->Add).

Change-Id: I7a8801de4b69796f3291a9e3a5896a055859121f