Adds glance snapshot to the controller node
Snapshot 'glance_installed' will contain the following packages

    1. glance

Glance will be installed and configured, also configures keystone
to authenticate glance. Adds Cirros_x86_64 to glance.

This snapshot can be used to obtain stable running version of
controller node and can be used to revert back to this point in case
the students break the cluster.

Change-Id: If97fdb04cfd1495e85a4c2cc44328c555caff0f6