labs: fix race in preseed script
When using the VirtualBox additions, there is a chance that after the
vboxadd service has started, the shared directory is not yet fully
available. In that case, activate_autostart.sh is never run (although
it may appear in time to be deleted).

This patch makes the script from preseed wait for activate_autostart.sh
to become available before it tries to execute it.

Change-Id: I1c2f3ef2f4301f36efec8a70b811ffe97dbaa779