labs: dnsmasq config: /etc/hosts, DNS server, logging
Configure dnsmasq to use the Google Public DNS and the node VM's
/etc/hosts for resolving hostnames (lets instance VM resolve the names
of node VMs, e.g. ping controller-api).

This configuration is in effect as long as the instance VMs use dnsmasq
for DNS name resolution (they do not if the neutron subnet has
dns_nameservers configured).

Since we turn on verbose logging in neutron scripts on the network node
VM, enable logging for dnsmasq, too.

Change-Id: I4bf470ddd87034fc0dcf29b612860ccda057f5e0