labs: Creates insecure ssh key
This patch only adds insecure ssh keys which will replace the vagarant
insecure ssh keys used by osbash.

We need to merge this patch before we can use these keys as the public
key needs to be downloaded into the kernel while installing the
operating system on all the nodes to enable ssh to the VMs.

This patch will be followed by another patch which will use they osbash
ssh key and remove the vagarant keys from the repository.

Change-Id: Ic3b6d7cbad1077a967a3396a21a6c4c91e3ae091