Add KVM provider (alternative to VirtualBox)
This patch adds KVM as an alternative provider of node VMs. KVM's main
advantages over VirtualBox are its speed (on platforms that have it) and
its availability on CI infrastructure.

As of now, not all osbash features are supported with KVM. For instance,
no snapshots are taken during the cluster build. The code also makes
assumptions, for instance about the availability of a default KVM
network.

Test with:

PROVIDER=kvm ./osbash.sh -b cluster

Change-Id: If88c6413913114beeab444559eccd31007b899c8