Auto-update ISO_URL
Every time a new Ubuntu installation ISO image is released, the old ISO
image is removed and the labs scripts can no longer build new base disks
until ISO_URL and ISO_MD5 are updated.

With this patch, the labs scripts find and use the new Ubuntu ISO image
if the old one is neither in the local IMG_DIR directory nor on the server.

Change-Id: I41483f2c63a5cda36a5101e824d87e85b7de0161