Add option to skip snapshot cycles
This patch adds a new option for ignoring the snapshot_cycle command.
Instead of shutting down the machine, taking a snapshot and rebooting it
again, the next script in the list is immediately executed. Only
snapshots created with the snapshot_cycle command are affected.

This speeds up the build time for the cluster considerably.

The patch also starts the network interfaces in init_xxx_node to have
them ready for the next script without having to reboot the machine.

Change-Id: I12a11144113a1fcde60a64a525e2217f4a758577