Make check_md5 function generic
The check_md5 function is generic and belongs into functions-host.sh,
even though it is currently only used in virtualbox-install_base.sh.

This patch also touches up the status and error messages to make them
generic. The function now aborts if the file it is supposed to check
does not exist. In addition, it no longer removes a file if the MD5 sum
does not match, which gives the user a chance to investigate what
happened.

Change-Id: If0f3d8fd69645638ecc659d0d676a425cd8b3167