Simplify tox.ini bashate call
The find arguments tox uses to identify files for bashate to check miss
some files that should be checked and include several files that should
not be checked.

For users and developers, it is not obvious if their files are going to
be checked or not because layers of criteria are stacked on top of
each other.

I propose a simple rule: outside of dot dirs and the doc folder, files
whose names match *.sh are checked, the rest is not. This was the
initial plan, too, and we renamed several files because of that. For
the few library files that we missed at the time, the proper fix is to
rename them, too.

Change-Id: If4d35b4e80fed54c69cbe4981a87b0c2a7338b9b