Fix shutdown on Windows
The shutdown.sh script used for the snapshot cycle is currently a
symbolic link. Depending on how the symbolic link is transferred
to Windows, it may not work on that operating system.

In order to resolve this issue, this patch removes the symbolic
link and makes the snapshot_cycle use the regular file instead.

Change-Id: I094fd7f7772fb89883a77d81c3b92f431c70ec07