Use qcow2 in KVM: compressed, copy-on-write images
In KVM, we have used raw disk images which were cloned for the node VMs.
Cloning is slow and the disk images use a lot of space.

This patch uses qcow2 disks. If virt-sparsify is available, the base
disk image is made sparse and compressed.

The node VM images are now copy-on-write images with the basedisk as
a backing image.

Taken together, these changes cut down the space used by all disks from
17 GB to 4 GB, and the time spend cloning the disks is saved.

Change-Id: I35a2234d11159c05c7ab844b49eec737d7e41496