Shut down running labs VMs before building cluster
Running cluster VMs from an earlier cluster build have the potential to
influence nodes (especially the controller node) while they are being
built.

To prevent that, this patch sends an ACPI powerbutton signal to all VMs
in the group VM_GROUP.

For KVM, we use the VM description's title field to store the group
name.

Change-Id: Ibe0a38a407c5e569b9a792faa7cf8b184f35022a