Fix iniset_no_section
The iniset_no_section function replaces strings if the option name
occurs anywhere in a line of the configuration file. That includes
options that are commented out and even completely unrelated lines that
just happen to contain an option name somewhere.

This patch uses the same logic as functions-common-devstack from which
the function was copied to begin with.

Change-Id: I32dae18173b2b0ca27ba144a19ddf535ced4ba21