KVM: remember IP addresses of nodes
In the current code, we get the IP address of a VMs first network
interface once via arp; for subsequent requests, we log into the VM via
the management network interface to get the information.

This requires hard-coded assumptions in the host-side code: each VM must
have an active, statically configured interface in a network called
"mgmt".

With this patch, we just store the information we get from arp the first
time around in a text file and retrieve it when needed. We still have to
go through arp only once, but we can drop the assumptions mentioned
above.

Change-Id: Ieb317638b4ddb28d2a54571ad6536efd4511bef7