Fix IP address changes with KVM
With KVM, the current code does not allow running scripts on a node
after another node has already been set up, because with the KVM
backend, the node setup is the only time during the cluster build when
the ssh connection information is updated.

The issue is fixed by this minimal patch.

Change-Id: I7d988db9ff8d8ff5895f0921e2c10f317136addd