Prevent snapshot name collisions
Skip snapshot taking if a snapshot with the given name already exists.

This is mostly useful when rebuilding (the remainder of) a cluster from
a snapshot. We do not want another series of snapshots for every test
run.

To that end, call the new vm_conditional_snapshot from generic code and
use new functions (vm_snapshot_list, vm_snapshot_exists) in back end
specific libraries to do the work.

Change-Id: If1de62bdf4272e3c9e1a198a104b1c2096ac01f7