Fix iptables to allow new connections between KVM networks
The network interfaces we created until Liberty differed from the
install-guide in that the node network interfaces in the public network
had an IP address, whereas the install-guide listed them as unnumbered
interfaces.

For Liberty, we are making the interfaces in the public network
unnumbered, too. With this change, the management network and
our pseudo-public network (203.0.113.0/24) can no longer establish
connections with each other, because the way libvirt configures iptables
only forwards established and related connections between these networks,
but not new ones.

Our scripts are supposted to leave the host system alone, but since
libvirt already messes with the host system's iptables and the training
cluster will not work without patching up libvirt's configuration work,
this change seems justified.

The patch modifies libvirt's forwarding rule for virbr interfaces (other
than virbr0 which is used for the default netowrk) and removes the
modifications before destroying the networks so libvirt recognizes the
rule again as its own.

One downside is that this patch prevents libvirt from removing these
forwarding rules if it is called in some other context.

Change-Id: I1e4134cb04e83c3ff69f89f8a1bc4b6ec586e6cd