Fix setting of KVM domain description
With KVM, the description strings we use to identify VMs belonging to
the training cluster are not written to disk and disappear as soon as
VM shuts down.

This patch adds "--config" and "--live" arguments to the virsh desc
call to fix this.

Also, in order to successfully write the description strings on a
freshly created domain, it is not enough to wait for it to appear in
the domain list: doing so leaves a race open that often results in an
error message about trying to change the persistent config of a transient
domain. Waiting for the domain to be "running" fixes that problem.

Change-Id: I63dc8befd436e8ef7e4bbe4fcc36cf4454298a3e