Get admin credentials for service list
Calling "openstack compute service list" requires admin credentials.
The test script does not provide them and the call fails. Because the
second call checked for the absence of the string "down", the script
continued, seemingly successful.

This patch adds the required credentials and changes the second test to
wait for the existence of "| up " instead.

Change-Id: I20da80a000a94b47b946a939f837913909540ab5