Check for other VMs before launching test VM
If an instance VM is already running, there are not enough resources to
launch a test VM (with the default RAM allocation). Such a VM could, for
instance, be a left over VM from a heat test.

With this patch, the test script aborts if any VMs are left running
after it tried to delete VMs created by earlier runs of the script.

Change-Id: I7ab09c3d0b9eca1f65df9682fe2f9b5b7c7cb027