Add option for disabling verbosity
In several places, the install-guides mentions how the user can choose
to enable verbose logging for OpenStack services.

So far, our scripts enable verbose logging for all services. Of course,
it can be turned off by changing the scripts in (as of now) 15 places,
but that is rather tedious.

With this patch, verbose logging can be turned off in a configuration
file or by setting an environment variable.

Change-Id: I6eba864756d1d9a13defb1118deaea42f3df8a45