Split logs by installation phase
This patch provides a mechanism for splitting server log files from
nodes into chunks that correspond to time periods (by default, roughly
from the begin of every script start to the begin of the next script
start).

In order to minimize the resources used for this (compute and storage),
the log_point function uses ls(1) to record the size of interesting
log files. These ls-snapshots are retrieved at the end of a
repeat-test run along with the long files.

A new tool, log_snapshot_split, uses that information to create a
directory for each snapshot. For all observed log files, each directory
contain the files and lines written since the previous ls-snapshot.

The whole tool chain can be called manually but is automatically used
by repeat-test.sh.

Change-Id: I809e485844b343bc1a39302396ec0f68a6801371