Wait for all KVM console types
The --wait option to make virt-install wait for the VM to complete the
install is necessary whenever we don't have a controlling console. That
is not only the case for installations started with the "-g headless"
option, but also for the other methods if the console viewer cannot be
opened, for instance, because the DISPLAY variable is not set. Without
the option, the VM does not get rebooted and must be manually started
for the installation process to continue.

This patch adds the --wait option for all KVM console types.

Change-Id: I13a40ec31e4d90afb83ba766ec8f66a0e01e63a5