Use arp as fallback for KVM IP database
With KVM, we use a text database that links hostnames and IP addresses.
Since the database is destroyed at the start of each osbash run and the
entries are added (by node_to_ip) whenever a new node is created, we can
safely assume that the entries exist and are correct when we needed them
at a later point (in ssh_env_for_node).

The assumption would not hold, however, if we were restarting a build
from a snapshot, which the KVM backend is due to support soon.

To cover this and possibly other cases, this patch makes ssh_env_for_node
call node_to_ip if an entry or the database file are missing.

Change-Id: Ia6b2fd2d931a20b4f084292e617894cb35900978