Mitaka pre-download update for swift
This patch updates the swift section of apt_pre-download for Mitaka.
We don't have scripts for object storage yet, but it allows users to
install swift on their own and it makes it easier to work on the
swift scripts if the same basedisk can be used. The impact on the
basedisk size is negligible.

Change-Id: Iefbf44bd2c37cb66005560998c91a12d72a66dc6