Replace loop device with real block device
This patch adds a new configuration option SECOND_DISK_SIZE for node VM
configuration files. If set, a block device of the given size is created
and attached to the VM (typically to become /dev/sdb).

This new option is used by setup_cinder_volumes, replacing the previous
method which used loop devices. The new client-side code now matches the
install-guide in this regard and is also easier to read.

Change-Id: I16b6a569d19676d7fff07592bd41bc71fa7c3ae2