Fix bridge setup when using Virtualbox additions
The environment for client-side scripts when called by osbashauto is not
exactly the same as when they are executed over ssh. It is the same user
(osbash), but the PATH differs and does not contain /sbin.

This patch adds the full path for brctl so the program is found on
Windows hosts (or other hosts using osbashauto) as well.

Change-Id: I23790ad4c5ac1b2f5e6afd2ee2e786abc7a51797
(cherry picked from commit b11bea604ec240a1a16076f509b62af1d70d7a1d)