Add preseed-ssh-v3.cfg
The new pressed file, preseed-ssh-v3.cfg, selects /dev/sda as the disk
to partition. This is not required for current basedisk builds (because
there is only ever one disk), but it is necessary for installation with
multiple disks (for instance PXE booting on nodes with a second disk
for storage).

The change should not break current installations, but adding a new
file keeps the old files around in case there is ever any doubt.

The new file also documents how to halt after installation (something
we might want to do in the future).

Change-Id: If291d1f223e56d7f76e7602a0db27687e82676f4