Complete proxy support
The current proxy support in osbash can only be used to speed up the
process. This patch improves the existing code which allows osbash to
build the basedisk in environments where an HTTP proxy is the only
Internet connection available.

To use it, set the VM_PROXY variable in config/localrc.

The patch adds a new preseed file that does not require an HTTPS
connection during the operating system installation.

Partial-Bug: #1602272
Change-Id: I89a7539292bd034c1b87013e8b2bc68777726601