Make shutdown command work on Windows
The recently introduced shutdown command did not make it into the
Windows batch files when calling osbash.sh only with the "-w" option
because in that case, vm_is_running would always return false.

This patch drops that check in vm_acpi_shutdown entirely. It is redundant
because every call to vm_acpi_shutdown is made from a location where the
VM is known to run (of course the VM can be stopped in the meantime,
but the removed check just made the window for that race somewhat
smaller without actually eliminating it).

Change-Id: Ia46705b2e9e78d9746a0c14685d724c53bfcdb47