Do not try to check for existing snapshots on Windows
There is no code to check for existing snapshots on Windows. Listing
them for that purpose just produces an error condition if the snapshot
does not exist (the common case) and the build aborts.

This patch removes the snapshot listing call from the Windows batch
files; it fixes a regression introduced with commit
3678ef2fa9893468604411ff697361bbd48d1e73.

Closes-Bug: #1606766
Change-Id: I221cabdf5e9948dda81cefdff7c8f3b3ab96817c