Collect kernel log buffer, too
When collecting information from the node VMs after tests, get the
content of the kernel ring buffer as well.

When investigating issues during the cluster build, we often find
warnings and errors in the kernel ring buffer. By saving the kernel log
for all runs, we get a baseline of normal behavior.

Change-Id: Ic11ccd79f69772076a844d65fe1f958d69ed2cf0