Add preseed file for Ubuntu 16.04 LTS
This patch adds a new preseed file that can be used to activate
osbashauto (our mechanism for autostarting queued scripts in the VMs on
Windows) in Ubuntu 16.04 LTS (reference platform for OpenStack's Newton
release). It should also work with other distributions that use systemd.

The changes that make use of this preseed file will come in a separate
patch so they can be tested (which requires the preseed file to be
already present and merged).

Change-Id: I78878821a43c5bc96b6a6476e60996d9000a25c4