Wait for VirtualBox share to become available
If we try to mount to the VirtualBox shared directory before it becomes
available, we get an error back and the installation aborts.

This patches fixes this race by making osbashauto try to mount the
share until it succeeds.

The patch also saves a log file to track the results.

Change-Id: Ic145aa7513292a88a9591955863a6b9bdcba866d
(cherry picked from commit e229aea8abf35c2db119c21718d4064feb8390a0)