Wait longer for installer to be ready
After starting the VM for building the basedisk, we currently wait for
5 seconds for the installer to be ready before we start sending our
prepared keyboard input.

This patch increases the waiting time to 10 seconds.

Rationale: 5 seconds is just about enough. On a fairly typical laptop,
it becomes a race, on a slower laptop, chances are the installer won't
be ready by the time we start giving it commands. If we don't wait long
enough, the installation will fail. Ideally, we would have a different
method of detecting that the installer is ready. For now, we just have to
try and err on the side of caution.

Change-Id: I6dfb96d758c1facbc0b4db570b80c36132fa561e