Do not get stuck on broken ssh connection
If the ssh server is not responding (usually due to a problem with the
target VM's kernel), the scripts wait indefinitely.

This patch makes ssh time out if the connection fails despite an open
ssh port.

One way of testing the behavior of an unresponsive ssh server without
making the VM's kernel fail is to stop the ssh server and start "netcat
-l 22" instead. It will reply on port 22 but completely fail to negotiate
an ssh session.

Change-Id: Ic2287eb3636418449ece464c1cf7fd6c6380ddf8