Handle ssh problems in get_node_logs.sh
This patch improves the behavior of the helper script, get_node_logs.sh,
when a target VM's ssh server is not unresponsive.

First, it prints a message to inform the user about the problem.

Second, it skips the code for splitting log files if an unresponsive ssh
server prevented the VM's log files from being retrieved.

Change-Id: I56bf7a4e229ca72226bfc308f30a869c8038fe14