Wait for VMs to stop with KVM, too
Since we introduced PXE booting for VirtualBox clusters with Change-Id
Ia46705b2e9e78d9746a0c14685d724c53bfcdb47, stop_running_cluster_vms
waits for the VMs to stop before returning. This change of behavior was
required to solve a race when later in the code, a VM was categorized
as running when it was in fact in the process of shutting down.

To be on the safe side and also for the sake of consistency between
VirtualBox and KVM, this patch makes stop_running_cluster_vms wait
for the VMs to stop when using KVM as well.

Change-Id: Ia12d3c08c702c4883ce8eb82fd02e32973d8eafe