Slow down keyboard input on Windows
The Windows batch files do not have to turn each character of the
boot parameters into a keyboard scancode before sending them to the
hypervisor. On some Windows hosts, the keycodes are sent too fast,
overruning the keyboard buffer. Characters are lost and the build fails.

This patch adds a few seconds to the basedisk build on Windows to avoid
this problem. Whenever a number of characters are sent, we sleep for a
second, giving the keyboard buffer time to be read.

Change-Id: I8fd952c020bd021538dab81acd15bc5908498cdf