Fix openstack subnet calls in launch_instance_private_net
The previous patch tries to set the DNS nameserver for a subnet
even though that name server is already present after a standard
cluster build. The openstack CLI client throws an error in such
a situation and the test script fails.

With this patch, we do not try to set a nameserver if one is
already configured.

Change-Id: I6ef36f118184092a57585eb4c63dd77bc3061788