Fix test status indication
When test scripts fail, repeat-test incorrectly logs "Test passed".

This patch makes test-once pass on the return status it got from the
test script. This is not sufficient for repeat-test to receive the error
(a pipe to tee eats the return status). For now, this patch uses a
simple solution: we grep the test log file for the actual return status.

Change-Id: I22ee768c4634e25e2a3f9154cf9a3cd69c2e0250