Fix copied openrc files
Our openrc files load some values from configuration files. For the
copies in the home directory, the relative paths are wrong.

This patch replaces the variables with constants which a) ensures
that the file works regardless of its path and b) is easier to
understand for the user.

We also replace the "controller" hostname in the openrc files with
the equivalent IP address so the openrc files works also on the
user's host machine.

Change-Id: I237f2449f8d2d5bd52166b8267b384900935b133