Fix use of absolute paths with log_snapshot_split
If absolute paths are passed to log_snapshot_split.py (that is, if the
"ls -l" lines contain absolute paths), then the result files will be
written to these paths which usually means that the original log files
are overwritten.

The script has always been used with relative paths in training-labs,
but any use outside of the project could lead to unexpected outcomes.
To prevent such surprises, this patch makes sure that absolute paths,
too, have the resulting files put into the resultdir supplied by the
user.

In addition, move the creation of result_path down to make the code
easier to understand.

Change-Id: I1a97e827151beb1fb0ce2b8c3308ad0b5b9b5c8c