New preseed files for Ubuntu 16.04.2
With Ubuntu 16.04.2, passing "minimal" to "tasksel/first" causes an
error.

The new preseed files use "standard system utilities" instead.

Fixing the existing files instead of creating new ones would fix
existing 16.04 installations that will keep pointing to the old
preseed files, but it might break 14.04 installations with no
way of fixing them.

Therefore, leaving the old preseed files for 14.04 based installations
and creating new ones for 16.04 based is the safer choice.

Note that this means we have to fix the preseed names in the Newton
branch.

Change-Id: I113c2adaecc79c54677c329aff4e5c3e2e0b3987