Change default to skip snapshot cycles
The snapshots we take after each installation step are mostly used for
debugging. Regular users seem to prefer build speed. This patch changes
the default to skipping the snapshot cycles (and quite a few reboots).

The option '-q' is now ignored. A new option '-e' to re-enable the
snapshot cycles has been added to st.py and repeat-test.sh.

Change-Id: I054cad556a11963f7df8ba65af85bde843a23656