Change preseed files to fix basedisk building with 16.04.2
This patch is a follow-up to the previous patches that tried to
fix basedisk building with the new minor version of the Ubuntu
ISO image, 16.04.2.

The previous fix resulted in a build that sometimes worked and
sometimes didn't. This patch moves all software installation
to the end of the installation.

The patch adds the openssh configuration to the vbadd preseed flavor
and makes the all preseed flavor an exact copy of vbadd. In time,
one of them can be removed.

Change-Id: I0cb0ae64d348c70754a4c1c146298164366202b7