Fix mount.vboxsf symlink (again)
With some versions of the VirtualBox guest additions, the symlink,
/sbin/mount.vboxsf, is broken. We have as_root_fix_mount_vboxsf_link
because of that. It creates /usr/lib/VBoxGuestAdditions if it does not
exist.

A new bug observed with the VirtualBox guest additions 5.1.20 causes the
symlink to point into a sub directory of where the binary actually
resides, resulting in the basedisk hanging after activate_autostart has
finished (when osbashauto tries to mount the shared directory).

This changeset makes it possible to build the basedisk in that case, as
well.

The bug has been fixed upstream with VirtualBox guest additions 5.1.22 [1],
but considering that this is at least the third time this particular symlink
has been broken, a generic fix like this one seems to be worth having.

[1] https://www.virtualbox.org/ticket/16670

backport: ocata newton

Change-Id: I2c8776b5da4e5adb87c9ec701d798c18ce56763b