Stop cluster VMs when building basedisk
A new basedisk can be built while a training-cluster is running (as long
as the new basedisk does not replace the basedisk used in the cluster).
However, on systems with 8 GB RAM there is not enough memory to keep
controller, compute and base VM running concurrently which can lead to
the host system becoming unresponsive.

With this patch, all running cluster VMs are shut down before a basedisk
is built.

Change-Id: Ic3cbd3ec1ab75e74ae67added4046f9d56c4b3b3