Set tmp directory for virt-sparsify
virt-sparsify uses about 10 GB temporary work space. The default (/tmp)
is often too small (especially if it is a RAM disk) and filled up by
virt-sparsify which hangs when it runs out of space.

With this changeset, we use the pool_dir as a tmp directory. On typical
systems, this means that instead of /tmp, the root partition needs to
have an extra 10 GB of space left, which should be much more likely to
work.

Change-Id: I4bf66474d6338c0e2e8eb76126ee496b13bc3c46