Clean up openrc copies
When writing the openrc copies that we put into the osbash user's home
directory for the users to find, we pick up superfluous comment lines
from the source file that lack the appropriate context in the resulting
openrc file.

This patch changes the logic so only the lines we need (starting with
"export") are kept. This also gets us closer to the install-guide.

Change-Id: If60479adbfecf9713b9981529fb369b13a5c89ad