Update Python code for Windows batch files
The osbash.sh code to produce the Windows batch files has seen several
improvements that have not been ported to Python. This changeset brings
all these changes to st.py.

To test the changes, it is sufficient to verify that the Windows batch
files generated by osbash.sh and st.py are now equivalent again.

Change-Id: I8514635e5b97f1b9388d532f32c7d2b2d35aecbf