Dynamic interface names for Windows batch scripts
On some systems, Windows VirtualBox starts numbering its host-only
interfaces (e.g., VirtualBox Host-Only Ethernet Adapter #2) at higher
numbers even if no such interfaces are already present.

In these cases, our scripts fail because they assume that the interfaces
to use are named
- VirtualBox Host-Only Ethernet Adapter
- VirtualBox Host-Only Ethernet Adapter #2
- VirtualBox Host-Only Ethernet Adapter #3
- [...]

With this changeset, create_hostnet write the names of the host-only
interfaces it creates to a configuration file which is used by the
the script that builds the cluster.

This will work regardless of the interface names returned by VBoxManage.

In order to write the correct variable names in all places, we use the
Unix-type interface names (e.g., vboxnet0) as Windows variable names
(e.g., %vboxnet0%).

Change-Id: Ieb67e646e8eaa96a57ff507283c74307350847e2