More RAM for basedisk build
Our default (512 MB RAM) is no longer sufficient to build the basedisk.
We have seen multiple out-of-memory errors. Building the cluster takes a
lot more RAM anyway, so we might as well double the RAM allocation for
the basedisk build.

This changeset is a minimal problem fix which also shows that the code
is inconsistent. A refactoring patch will follow later.

Change-Id: I55f52c50e4cde8583b8fea7a0942bfd9e8939f72