Remove missing swift configuration files
The Mitaka version of the swift configuration files have been removed
from the swift repo because Mitaka is end of life. This prevents the
base disk from building.

We could update the URLs to Ocata, but then our Ocata release would stop
working as soon as Ocata is EOL just because these files will be removed
as well. That is entirely unnecessary because right now, swift is
unsupported and the configuration files are not needed. For that reason,
just remove the configuration files until we need them.

Change-Id: I0739feec073d76cc62ade92d21b4e6bb8d1694f3
(cherry picked from commit bb26a284be8a9115fac798d830efa79f11511fe8)