Fix test_horizon for curl
The horizon test script did not work when using curl. It went unnoticed
because we always use wget for testing.

The optional arguments POST_ARG and REFERER_ARG were uninitialized
(an error with nounset) and passed as empty string arguments if set
to an empty string.

This patch fixes the code path for curl.

Change-Id: I7baeb1b409d9bbd90eb57493f5428a7b9f404644