Sync changes in scripts dir to node VMs
Until now, the scripts directory has not been synced to the node VMs
because each script was copied there prior to execution. This method
does not work if a script needs additional files (e.g., configuration)
that reside in the scripts directory. This is what we plan to do for
some optional services which will come in separate sub-directories
such as scripts/ubuntu/tacker.

Therefore, this changeset ensures that the scripts directory is copied
to the target node VM before script execution begins.

Change-Id: Iee4ed2fdc99045e779575a68a5c19c35895bdbd8