Drop LBaaS
Our LBaaS v1 scripts stopped working with the Kilo release, and they
have never been migrated to LBaaS v2 which is due to become deprecated
soon [1][2].

It is time to retire the scripts for this service. LBaaS can be
reintroduced when we have someone to write a script for Octavia.

[1] http://specs.openstack.org/openstack/neutron-specs/specs/newton/kill-neutron-lbaas.html
[2] https://wiki.openstack.org/wiki/Neutron/LBaaS/Deprecation

Change-Id: I8872dd4298dfc7fbfc4e53a403abd818b3ef47a8