Fix KEY_SHIFT keycode for libvirt
In 2017, libvirt changed the way they handle keycodes (presumably when
they switched to the keycodemapdb repo). They dropped the KEY_SHIFT
keycode which used to be mapped to KEY_LEFTSHIFT, resulting in an
error "invalid keycode: 'KEY_SHIFT'" when building the basedisk with
KVM.

Fortunately, KEY_LEFTSHIFT existed before, so we can just use that and
it will work on both new and old versions of libvirt.

Change-Id: Iaddd068a6ee19d8c7f027558dcf3d3efea26dbbb