[6/7] Ignore exit status for nova-status upgrade check
Starting with OpenStack Rocky, "nova-status upgrade check" returns an
apparently erroneous warning like this:

  Check: Resource Providers
  Result: Warning
  Details: There are no compute resource providers in the Placement
    service but there are 1 compute nodes in the deployment.
    This means no compute nodes are reporting into the
    Placement service and need to be upgraded and/or fixed.
    See
    https://docs.openstack.org/nova/latest/user/placement.html
    for more details.

With this changeset, an exit status of 1 coming from this command is no
longer treated as a reason for stopping with an error.

The ability of the training cluster to allocate resources and launch
VMs is not affected by this.

Change-Id: Ifb88cc3fe66314bfbf7962fda9911da884b0b1bc