Revert "Python 3 compatibility: convert raw_input to input"
The reverted patch introduced a library that is not part of a standard
Python install. Up to that point, we carefully avoided such dependencies
in the code, because it makes the software more difficult to install and
use. The patch provides no actual benefit (it was just a trivial first
step towards Python 3 support).

This reverts commit 03a35690a2f2067784188481882c079dbf4eb42b.

Change-Id: I11f823bf974931b5eafa5c74a003b0506543d37d