Fix disk multiattach with VirtualBox 6
Starting with version 6.0, the behavior of VirtualBox with regards to
disk multiattach changed.

The result are error messages like:
  VBoxManage: error: Cannot change type for medium '<base_disk_path>':
  the media type 'MultiAttach' can only be used on media registered
  with a machine that was created with VirtualBox 4.0 or later

The new code should work for both VirtualBox 6 and older versions.

The workaround suggests that we may not be using the VirtualBox volumes
the way they are meant to be used, but with scant documentation out
there rewriting the volume logic may result in no improvement at all,
so let's leave it at that for the time being.

backport: rocky queens pike ocata

Closes-Bug: 1817584

Change-Id: I9307d2e0f077539c118f540a9b0a4358e4f3b459
(cherry picked from commit e4dec38469acdac6f9e408ac8612e709405cd998)